Use reflection when roughness is zero in microfacet brdfs.
