+++ src/appleseed/renderer/modeling/bsdf/glossybrdf.cpp
@@ -33,7 +33,9 @@
 #include "renderer/kernel/lighting/scatteringmode.h"
 #include "renderer/modeling/bsdf/bsdf.h"
 #include "renderer/modeling/bsdf/bsdfwrapper.h"
+#include "renderer/modeling/bsdf/fresnel.h"
 #include "renderer/modeling/bsdf/microfacethelper.h"
+#include "renderer/modeling/bsdf/specularhelper.h"
 #include "renderer/utility/messagecontext.h"
 #include "renderer/utility/paramarray.h"
 
@@ -91,7 +93,7 @@ namespace
         GlossyBRDFImpl(
             const char*             name,
             const ParamArray&       params)
-          : BSDF(name, Reflective, ScatteringMode::Glossy, params)
+          : BSDF(name, Reflective, ScatteringMode::Glossy | ScatteringMode::Specular, params)
         {
             m_inputs.declare("reflectance", InputFormatSpectralReflectance);
             m_inputs.declare("reflectance_multiplier", InputFormatFloat, "1.0");
@@ -158,11 +160,24 @@ namespace
             BSDFSample&             sample) const APPLESEED_OVERRIDE
         {
             const Vector3f& n = sample.m_shading_basis.get_normal();
-            const float cos_on = std::min(dot(sample.m_outgoing.get_value(), n), 1.0f);
+            const Vector3f& outgoing = sample.m_outgoing.get_value();
+            const float cos_on = std::min(dot(outgoing, n), 1.0f);
             if (cos_on < 0.0f)
                 return;
 
             const InputValues* values = reinterpret_cast<const InputValues*>(data);
+
+            const FresnelDielectricFun<float> f(
+                values->m_reflectance,
+                values->m_reflectance_multiplier,
+                values->m_precomputed.m_outside_ior / values->m_ior);
+
+            // If roughness is zero use reflection.
+            if (values->m_roughness == 0.0f)
+            {
+                SpecularBRDFHelper::sample(f, sample);
+                return;
+            }
 
             float alpha_x, alpha_y;
             microfacet_alpha_from_roughness(
@@ -170,11 +185,6 @@ namespace
                 values->m_anisotropic,
                 alpha_x,
                 alpha_y);
-
-            const FresnelDielectricFun<float> f(
-                values->m_reflectance,
-                values->m_reflectance_multiplier,
-                values->m_precomputed.m_outside_ior / values->m_ior);
 
             if (m_mdf == GGX)
             {
