+++ src/appleseed/renderer/modeling/bsdf/specularbrdf.cpp
@@ -34,6 +34,7 @@
 #include "renderer/kernel/lighting/scatteringmode.h"
 #include "renderer/modeling/bsdf/bsdf.h"
 #include "renderer/modeling/bsdf/bsdfwrapper.h"
+#include "renderer/modeling/bsdf/specularhelper.h"
 
 // appleseed.foundation headers.
 #include "foundation/math/basis.h"
@@ -49,6 +50,33 @@ namespace renderer
 
 namespace
 {
+    template <typename T>
+    class NoFresnelFun
+    {
+      public:
+        NoFresnelFun(
+            const Spectrum& reflectance,
+            const T         reflectance_multiplier)
+          : m_reflectance(reflectance)
+          , m_reflectance_multiplier(reflectance_multiplier)
+        {
+        }
+
+        void operator()(
+            const foundation::Vector<T, 3>& o,
+            const foundation::Vector<T, 3>& h,
+            const foundation::Vector<T, 3>& n,
+            Spectrum&                       value) const
+        {
+            value = m_reflectance;
+            value *= static_cast<float>(m_reflectance_multiplier);
+        }
+
+      private:
+        const Spectrum& m_reflectance;
+        const T         m_reflectance_multiplier;
+    };
+
     //
     // Specular BRDF.
     //
@@ -91,30 +119,13 @@ namespace
             if (cos_on < 0.0f)
                 return;
 
-            // Compute the incoming direction.
-            const Vector3f incoming(
-                force_above_surface(
-                    reflect(sample.m_outgoing.get_value(), shading_normal),
-                    sample.m_geometric_normal));
-
-            // No reflection below the shading surface.
-            const float cos_in = dot(incoming, shading_normal);
-            if (cos_in < 0.0f)
-                return;
-
-            // Compute the BRDF value.
             const InputValues* values = static_cast<const InputValues*>(data);
-            sample.m_value = values->m_reflectance;
-            sample.m_value *= values->m_reflectance_multiplier / cos_in;
 
-            // The probability density of the sampled direction is the Dirac delta.
-            sample.m_probability = DiracDelta;
+            const NoFresnelFun<float> f(
+                values->m_reflectance,
+                values->m_reflectance_multiplier);
 
-            // Set the scattering mode.
-            sample.m_mode = ScatteringMode::Specular;
-
-            sample.m_incoming = Dual3f(incoming);
-            sample.compute_reflected_differentials();
+            SpecularBRDFHelper::sample(f, sample);
         }
 
         APPLESEED_FORCE_INLINE virtual float evaluate(
